+++ src/components/dropdown/types.ts
@@ -1,11 +1,14 @@
-import { TriggerType } from '../popper/types';
+import {
+  PopperPlacement,
+  TriggerType,
+} from '../popper/types';
 
 export interface DropdownBindProperties {
   value?: boolean;
   popperClass?: string;
   width?: number;
   trigger?: TriggerType;
-  placement?: string;
+  placement?: PopperPlacement;
   arrowVisible?: boolean;
   withBorder?: boolean;
   disabled?: boolean;